test(slack): split tests into per-concept modules
slack/tests.rs (607) split into six sibling files by concern: http_tests,
url_tests (parse/format/extract_target + msg helper), mention_tests,
constructor_tests, resolve_messages_tests, classify_tests. The shared
DummyBody deserialize target moves to slack.rs under #[cfg(test)] since
http_tests and constructor_tests both reference it. All files <=400 lines;
production logic unchanged; 449 tests pass.
